Implementing Require.js as a dependecy managment tool.
All the backbone classes are now laid out in a more MVC fashion
with views, collections, models, and templates. The entry point starts
in scripts/main.js then flows to scripts/app.js then to
scripts/router.js then to the appropriate classes.

From this commit things are not as they were before but every class
should be rendering. The next step is to make sure the css is set back
to what it was previously.